requirements.txt
hidapi
webcolors
